fix: move helper methods to a separate helper class
